Slit stores in their own files
