+++ server/.env.dist
@@ -1,9 +1,15 @@
+# Abuse mail configuration
 MAILGUN_API_KEY=api-key
 MAILGUN_DOMAIN=darkwire.io
 ABUSE_TO_EMAIL_ADDRESS=abuse@darkwire.io
 ABUSE_FROM_EMAIL_ADDRESS=Darkwire <no-reply@darkwire.io>
-REDIS_URL=redis://localhost:6379
+
 CLIENT_DIST_DIRECTORY='client/dist/path'
+
 ROOM_HASH_SECRET='some-uuid'
+
 SITE_URL=https://darkwire.io
-STORE_BACKEND='redis'
+
+# Store configuration
+STORE_BACKEND=redis
+STORE_HOST=redis://localhost:6379
