+++ server/src/store/Memory.js
@@ -1,43 +1,10 @@
-import Redis from 'redis';
-import bluebird from 'bluebird';
-
-/**
- * Redis store.
- */
-export class RedisStore {
-  constructor(redisUrl) {
-    bluebird.promisifyAll(Redis.RedisClient.prototype);
-    bluebird.promisifyAll(Redis.Multi.prototype);
-    this.redis = Redis.createClient(redisUrl);
-  }
-
-  get(key, field) {
-    return this.redis.hgetAsync(key, field);
-  }
-
-  getAll(key) {
-    return this.redis.hgetallAsync(key);
-  }
-
-  set(key, field, value) {
-    return this.redis.hsetAsync(key, field, value);
-  }
-
-  del(key, field) {
-    return this.hdelAsync(key, field);
-  }
-
-  inc(key, field, inc = 1) {
-    return this.redis.incrbyAsync(key, field, inc);
-  }
-}
-
 /**
  * Memory store more for testing purpose than production use.
  */
 export class MemoryStore {
   constructor() {
     this.store = {};
+    this.hasSocketAdapter = false;
   }
 
   async get(key, field) {
@@ -72,4 +39,7 @@ export class MemoryStore {
   async inc(key, field, inc = 1) {
     this.store[key][field] += inc;
   }
+
 }
+
+export default MemoryStore
